Re-structured more files
