+++ assets/images/badges/info.json
@@ -1,32 +1,32 @@
-[
-    {
+{
+    "25kSubs": {
         "id": "25kSubs",
         "name": "YouTu",
         "description": "So close to a YouTube rank! Suggested by ixNoah",
         "icon": "./img/25kSubs.png"
     },
-    {
+    "passingTime": {
         "id": "passingTime",
         "name": "Passing Time",
         "description": "Play on Hypixel for over 5 years",
         "icon": "./img/passingTime.png"
     },
-    {
+    "oink": {
         "id": "oink",
         "name": "Oink!",
         "description": "Be Technoblade",
         "icon": "./img/oink.png"
     },
-    {
+    "experienced": {
         "id": "experienced",
         "name": "Experienced",
         "description": "Claimed your 50 level reward",
         "icon": "./img/experienced.png"
     },
-    {
+    "noLife": {
         "id": "noLife",
         "name": "No Life",
         "description": "Claimed your 200 level reward",
         "icon": "./img/noLife.png"
     }
-]
+}
