+++ lib/hypixel/userProfile.mjs
@@ -1,5 +1,6 @@
 import fetch from 'node-fetch';
 import minecraftColors from '../colors/minecraft.mjs';
+import { getRank } from './playerData.mjs';
 
 // Import api key from .env
 import dotenv from 'dotenv';
@@ -44,103 +45,6 @@ async function getSkin(username) {
 
     // If the user doesn't exist or Mojang API is not working
     return 0;
-
-}
-
-// Get rank info on Hypixel
-function getRank(data) {
-
-    /*
-
-    How to get the rank:
-    
-    1. `packageRank` Permanent ranks before 2016 EULA update 
-    2. `newPackageRank` New ranks like "VIP_PLUS" after 2016 EULA update
-    3. `monthlyPackageRank` for monthly ranks like "MVP++" (aka. "SUPERSTAR")
-    4. `data.player.rank` for special ranks like "Admin", "Youtube" etc.
-    5. Hardcode if the player has special ranks e.g. "PIG+++", "Slothpixel"
-    6. `data.player.rankPlusColor` if the rank has "+" in it.
-
-    Thank you Hypixel :heart:
-
-    */
-
-    let rank = data.player.packageRank;
-    let newRank = data.player.newPackageRank;
-    let monthlyRank = data.player.monthlyPackageRank;
-    let specialRank = data.player.rank;
-
-    // If the user logged in at least once since 2016
-    if (newRank != null) {
-        rank = newRank;
-    }
-    
-    // If the user has a monthly rank
-    if (monthlyRank != null && monthlyRank != "NONE") {
-        rank = monthlyRank;
-    }
-    
-    // If the user has a special rank
-    if (specialRank != null && specialRank != "NONE") {
-        rank = specialRank;
-    }
-    
-    // Get full rank name
-    let formattedRank = "";
-
-    switch (rank) {
-        
-        case "VIP": formattedRank = "[VIP]"; break;
-        case "VIP_PLUS": formattedRank = "[VIP+]"; break;
-        case "MVP": formattedRank = "[MVP]"; break;
-        case "MVP_PLUS": formattedRank = "[MVP+]"; break;
-        case "SUPERSTAR": formattedRank = "[MVP++]"; break;
-        case "YOUTUBER": formattedRank = "[YOUTUBE]"; break;
-        case "ADMIN": formattedRank = "[ADMIN]"; break;
-        
-    }
-    
-    // Getting Technoblade's custom rank
-    if (data.player.uuid == "b876ec32e396476ba1158438d83c67d4") formattedRank = "[PIG+++]"
-    
-    // Get color of the rank
-    let color = "gray";
-    
-    switch (formattedRank) {
-        
-        case "[VIP]": color = "green"; break;
-        case "[VIP+]": color = "green"; break;
-        case "[MVP]": color = "aqua"; break;
-        case "[MVP+]": color = "aqua"; break;
-        case "[MVP++]": color = "gold"; break;
-        case "[YOUTUBE]": color = "red"; break;
-        case "[ADMIN]": color = "red"; break;
-        case "[OWNER]": color = "red"; break;
-        case "[PIG+++]": color = "light_purple"; break;
-        
-    }
-    
-    let plusColor = color;
-
-    // If the rank has a + in it
-    if ( formattedRank.includes("+") ) {
-
-        // Get the color of the plusses
-        plusColor = data.player.rankPlusColor.toLowerCase();
-
-        
-        // If the user has a pig rank
-        if ( formattedRank == "[PIG+++]" ) {
-            plusColor = "dark_purple";
-        }
-        
-    }
-
-    return {
-        rank: formattedRank,
-        color: color,
-        plusColor: plusColor
-    }
 
 }
 
@@ -230,6 +134,5 @@ async function getNickname(username) {
 
 export {
     getSkin,
-    getNickname,
-    getRank
+    getNickname
 };
